Refactor the fs interface (#269)

+++ internal/service/inodes/service.go
@@ -149,7 +149,7 @@ func (s *INodeService) CreateFile(ctx context.Context, cmd *CreateFileCmd) (*INo
 	return &inode, nil
 }
 
-func (s *INodeService) RegisterWrite(ctx context.Context, inode *INode, sizeWrite int, h hash.Hash) error {
+func (s *INodeService) RegisterWrite(ctx context.Context, inode *INode, sizeWrite int64, h hash.Hash) error {
 	inode.lastModifiedAt = s.clock.Now()
 	inode.size += uint64(sizeWrite)
 	inode.checksum = hex.EncodeToString(h.Sum(nil))
