+++ internal/service/inodes/service_test.go
@@ -555,7 +555,7 @@ func TestINodes(t *testing.T) {
 			"size":             ExampleAliceFile.size + uint64(n),
 		}).Return(nil).Once()
 
-		err = service.RegisterWrite(ctx, &ExampleAliceFile, n, hash)
+		err = service.RegisterWrite(ctx, &ExampleAliceFile, int64(n), hash)
 		assert.NoError(t, err)
 	})
